--- a/.github/workflows/py-hypertrace.yml
+++ b/.github/workflows/py-hypertrace.yml
@@ -7,6 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     container:
       image: pgbrodrick/isofit
+      options: --shm-size=2gb
 
     steps:
     - name: Checkout