Adds additional interpolation options, and changes the default to a 1-d ndsplines implementation.  This has a big speed effect (benchmark numbers to follow) when the number of dimensions in a LUT is moderate to high, with identical performance.  Alternately, a higher order spline can be used to decrease interpolation error.

When the new interpolation method is set to 'nd-1', the code below gives:

Original interpolation: 0.181365966796875 s
New interpolation: 0.07266783714294434 s

<img width="845" alt="Screen Shot 2022-02-17 at 1 44 00 PM" src="https://user-images.githubusercontent.com/23531204/154576790-59935cfe-3e3c-4e9c-80c0-98bda60bf40a.png">

When the new interpolation method is set to 'nd-2', the code below gives:

Original interpolation: 0.19042301177978516 s
New interpolation: 0.08439183235168457 s

<img width="878" alt="Screen Shot 2022-02-17 at 1 44 32 PM" src="https://user-images.githubusercontent.com/23531204/154576910-43061aab-c3fd-43d3-b7ad-a4fda1c6142a.png">

Speedups increase with more dimensions.  Benchmarking on scenes including inversions is still in progress, and careful checking is needed before acceptance.

Code used for comparison between methods:

```
import numpy as np
import matplotlib.pyplot as plt
from isofit.core.common import VectorInterpolator
import time
import matplotlib.gridspec as gridspec

def func(x):
    if x.shape[-1] == 2:
        #return np.matmul((np.cos(x[:,0]) * x[:,1]**1.1).reshape(-1,1), np.arange(10).reshape(1,-1))
        return np.matmul((x[:, 0]**0.9 * x[:,1]**1.1).reshape(-1, 1), np.ones(10).reshape(1, -1))
    if x.shape[-1] == 3:
        #return np.matmul((np.cos(x[:,0]) * x[:,1]**1.1 * np.sin(x[:,1]) ).reshape(-1,1), np.arange(10).reshape(1,-1))
        return (x[:,0] * x[:,1] ).reshape(-1,1)[:,:,np.newaxis] * np.ones(10)[np.newaxis,np.newaxis,:]


grid_input = [[1,5,9], [1, 3, 5]]
grid = np.stack(np.meshgrid(*grid_input, indexing='ij'),axis=-1)
grid = np.reshape(grid, (int(np.product(grid.shape)/grid.shape[-1]), grid.shape[-1]))
data_input = func(grid).reshape((3,3,10))
grid = np.stack(np.meshgrid(*grid_input, indexing='ij'),axis=-1)
lut_interp_types = np.array(['n','n'])

v_orig = VectorInterpolator(grid_input, data_input, lut_interp_types, 'rg')
v_new = VectorInterpolator(grid_input, data_input, lut_interp_types, 'nds-2')

input_test = np.random.random((1000,len(grid_input)))
for _n in range(len(grid_input)):
    input_test[:,_n] = input_test[:,_n] * (np.max(grid_input[_n]) - np.min(grid_input[_n])) + np.min(grid_input[_n])

res_orig = np.zeros((input_test.shape[0], data_input.shape[-1]))
res_new = np.zeros((input_test.shape[0], data_input.shape[-1]))

stime = time.time()
for _n in range(res_orig.shape[0]):
    res_orig[_n,:] = v_orig(input_test[_n,:])
print(f'Original interpolation: {time.time() - stime} s')
stime = time.time()
for _n in range(res_orig.shape[0]):
    res_new[_n,:] = v_new(input_test[_n,:])
print(f'New interpolation: {time.time() - stime} s')


X, Y = np.meshgrid(np.linspace(np.min(grid_input[0]), np.max(grid_input[0]), 10), np.linspace(np.min(grid_input[1]), np.max(grid_input[1]), 10))
Z_orig = np.zeros((X.shape[0],X.shape[1],res_new.shape[-1]))
Z_orig[...] = np.nan
Z_new = np.zeros((X.shape[0],X.shape[1],res_new.shape[-1]))
Z_new[...] = np.nan
Z_true = np.zeros((X.shape[0],X.shape[1],res_new.shape[-1]))
Z_true[...] = np.nan
for _x in range(X.shape[0]):
    for _y in range(X.shape[1]):
        Z_orig[_x,_y,:] = v_orig(np.array([X[_x,_y], Y[_x,_y]]))
        Z_new[_x,_y,:] = v_new(np.array([X[_x,_y], Y[_x,_y]]))
        Z_true[_x,_y,:] = func(np.array([X[_x,_y], Y[_x,_y]]).reshape(1,-1))

fig = plt.figure(figsize=(10,8))
gs = gridspec.GridSpec(nrows=2, ncols=2)

ax = fig.add_subplot(gs[0,0])
plt.pcolormesh(np.linspace(np.min(grid_input[0]), np.max(grid_input[0]), 10), np.linspace(np.min(grid_input[1]), np.max(grid_input[1]), 10),Z_orig[...,1],shading='nearest')
plt.colorbar()
plt.scatter(grid[...,0], grid[...,1], c='white')
plt.title('Original')

ax = fig.add_subplot(gs[0,1])
plt.pcolormesh(np.linspace(np.min(grid_input[0]), np.max(grid_input[0]), 10), np.linspace(np.min(grid_input[1]), np.max(grid_input[1]), 10),Z_new[...,1],shading='nearest')
plt.colorbar()
plt.scatter(grid[...,0], grid[...,1], c='white')
plt.title('New')


ax = fig.add_subplot(gs[1,1])
plt.pcolormesh(np.linspace(np.min(grid_input[0]), np.max(grid_input[0]), 10), np.linspace(np.min(grid_input[1]), np.max(grid_input[1]), 10),Z_new[...,1] - Z_true[...,1],shading='nearest', cmap='RdBu')
plt.colorbar()
plt.scatter(grid[...,0], grid[...,1], c='white')
plt.title('New Error')


ax = fig.add_subplot(gs[1,0])
plt.pcolormesh(np.linspace(np.min(grid_input[0]), np.max(grid_input[0]), 10), np.linspace(np.min(grid_input[1]), np.max(grid_input[1]), 10),Z_orig[...,1] - Z_true[...,1],shading='nearest', cmap='RdBu')
plt.colorbar()
plt.scatter(grid[...,0], grid[...,1], c='white')
plt.title('Orig Error')

plt.show()

```