@pgbrodrick 2022-04-06
@davidraythompson, after some debugging, the hypertrace workflow error is newly induced, best as I can tell from new resource changes to GHActions runs.  The workflow runs on at least one other machine. I vote we go ahead and merge this in, and put the hypertrace workflow update as a current bug.